services/auth-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt
mysqlclient>=2.1
django-cors-headers>=4.0
django-mysql
pika>=1.1
Pillow>=9.0